fix: add teacher-commnet-model subject-comment-model #107
