+++ fastapi/app/models/__init__.py
@@ -1,12 +1,13 @@
 from .base import BaseModelMixin  # isort:skip
 
 from .attend_subject import AttendSubject
-from .comment import SubjectComment, TeacherComment
 from .department import Department
 from .evaluation import Evaluation
 from .school import School
 from .score import Score
 from .subject import Subject
+from .subject_comment import SubjectComment
 from .teacher import Teacher
 from .term import SemesterEnum, Term
+from .teacher_comment import TeacherComment
 from .user import User
